CL: add APR breakdown to pool detail page (#2625)
* use apr from numia in CL pool card

* Add APRs to CL pool detail page

* fix wrap
