+++ packages/stores/src/derived-data/pool/concentrated.ts
@@ -15,6 +15,7 @@ import { OsmosisQueries } from "../../queries/store";
 import {
   ObservableQueryActiveGauges,
   ObservableQueryClPoolAvgAprs,
+  ObservableQueryPoolAprs,
   ObservableQueryPoolFeesMetrics,
   ObservableQueryPriceRangeAprs,
 } from "../../queries-external";
@@ -31,7 +32,7 @@ export class ObservableConcentratedPoolDetail {
       queryPoolFeeMetrics: ObservableQueryPoolFeesMetrics;
       queryActiveGauges: ObservableQueryActiveGauges;
       queryPriceRangeAprs: ObservableQueryPriceRangeAprs;
-      queryClPoolAvgAprs: ObservableQueryClPoolAvgAprs;
+      queryPoolAprs: ObservableQueryPoolAprs;
     },
     protected readonly accountStore: AccountStore,
     protected readonly priceStore: IPriceStore
@@ -82,12 +83,17 @@ export class ObservableConcentratedPoolDetail {
 
   @computed
   get swapFeeApr(): RatePretty {
-    const queryPool = this.osmosisQueries.queryPools.getPool(this.poolId);
-    if (!queryPool) return new RatePretty(0);
+    return (
+      this.externalQueries.queryPoolAprs.getForPool(this.poolId)?.swapFees ??
+      new RatePretty(0).ready(false)
+    );
+  }
 
-    return this.externalQueries.queryPoolFeeMetrics.get7dPoolFeeApr(
-      queryPool,
-      this.priceStore
+  @computed
+  get totalApr(): RatePretty {
+    return (
+      this.externalQueries.queryPoolAprs.getForPool(this.poolId)?.totalApr ??
+      new RatePretty(0).ready(false)
     );
   }
 
@@ -96,13 +102,6 @@ export class ObservableConcentratedPoolDetail {
       this.externalQueries.queryPriceRangeAprs
         .get(this.poolId)
         .apr?.inequalitySymbol(true) ?? new RatePretty(0).ready(false)
-    );
-  }
-
-  get avgApr(): RatePretty {
-    return (
-      this.externalQueries.queryClPoolAvgAprs.get(this.poolId).apr ??
-      new RatePretty(0).ready(false)
     );
   }
 
@@ -206,6 +205,7 @@ export class ObservableConcentratedPoolDetails extends HasMapStore<ObservableCon
       queryActiveGauges: ObservableQueryActiveGauges;
       queryPriceRangeAprs: ObservableQueryPriceRangeAprs;
       queryClPoolAvgAprs: ObservableQueryClPoolAvgAprs;
+      queryPoolAprs: ObservableQueryPoolAprs;
     },
     protected readonly accountStore: AccountStore,
     protected readonly priceStore: IPriceStore
