+++ packages/web/components/pool-detail/concentrated.tsx
@@ -29,6 +29,8 @@ import { formatPretty } from "~/utils/formatter";
 import { getNumberMagnitude } from "~/utils/number";
 import { removeQueryParam } from "~/utils/url";
 
+import { AprBreakdown } from "../cards/apr-breakdown";
+
 const ConcentratedLiquidityDepthChart = dynamic(
   () => import("~/components/chart/concentrated-liquidity-depth"),
   { ssr: false }
@@ -470,7 +472,7 @@ const UserAssetsAndExternalIncentives: FunctionComponent<{ poolId: string }> =
     const hasIncentives = concentratedPoolDetail.incentiveGauges.length > 0;
 
     return (
-      <div className="flex h-40 gap-4">
+      <div className="flex flex-wrap gap-4">
         <div className="flex shrink-0 items-center gap-8 rounded-[28px] bg-osmoverse-1000 px-8 py-7">
           <div className="flex h-full flex-col place-content-between">
             <span className="body2 text-osmoverse-300">
@@ -509,6 +511,11 @@ const UserAssetsAndExternalIncentives: FunctionComponent<{ poolId: string }> =
             ))}
           </div>
         </div>
+        <AprBreakdown
+          className="shrink-0 rounded-[28px] bg-osmoverse-1000"
+          poolId={poolId}
+        />
+
         {hasIncentives && (
           <div className="flex h-full w-full flex-col place-content-between items-center rounded-[28px] bg-osmoverse-1000 px-8 py-7">
             <span className="body2 mr-auto text-osmoverse-300">
